Add update.yml. Runs nightly to check for new commits.
If new commits are found, build_iAPS.yml is launched, and the target branch is synced with upstream.

Scheduled workflows in GitHub actions will be cancelled in the event of 60 days of no repository activity. A "keep alive" action will generate empty commits if there is no activity after 50 days to avoid inactivation of the scheduled workflow.

Added on: workflow_call, env variables and sync action to build_iAPS.yml.

Removed the "secrets" job", this can still be run from the identifiers and certificates workflows to validate repository secrets. 

Builds on macos-13 (currently in beta) and Xcode 14.3