Validate_secrets.yml: run every 5 minutes on schedule
Stress-testing the Validate Secrets workflow with Fastlane PR 21583 for errors like this:

'Authentication credentials are missing or invalid. - Provide a properly configured and signed bearer token, and make sure that it has not expired.'

I am comparing this with a similar stress-test in another repository to be able to evaluate the results.